tests: Don't use coproc because it's unreliable
The file descriptors are closed as soon as the command exits so any
delay before opening the coproc file descriptor causes the test to
fail.

Use a named fifo instead. The command will have to wait for us to
open the fifo before it can open or write to it.