Handle system calls that could return EINTR by retrying
The following functions are not documented as being interruptible but they
could block on filesystem access so it may be possible (macOS can return
EINTR on mkdir()):
  lseek()
  mkdtemp()

The close() function is interruptible but the file descriptor should still
have been closed (on Linux and FreeBSD but not on HP-UX). The behaviour is
undocumented on OpenBSD and NetBSD.

mkostemp() is not documented as being interruptible, but glibc calls open()
so it could return EINTR. This isn't a problem because this function is
only used before a command is executed so there are no signals to be
handled.

mkdtemp() calls mkdir() but it is only used before a command is executed.

waitpid() with WNOHANG is not interruptible on Linux, but POSIX allows it
to be.

This makes the stream based tests 100% reliable on Linux.