Makefile: Always use the default compiler
Using GCC fails badly on FreeBSD 11.2 because there are ABI issues that
cause it to fail to link to boost_program_options.

The downside to not specifying CC and CXX explicitly is that the default
compiler is GCC but Eclipse likes to specify clang so the compiler being
used can vary.