meson: Support compiling on GNU Hurd
Define appropriate values for PIPE_BUF and PATH_MAX.

It is impossible to use pathconf() to get a true maximum path size for the
target of a symbolic link before reading it.

It is not possible to use fpathconf() to get the maximum atomic write size
for a pipe/FIFO from a socket. A pipe is required.