stream_output: Configure standard output and standard error to be blocking
Non-blocking output streams cause all kinds of obscure problems and
should never be used with interactive programs. Make stdout and stderr
blocking if they're currently non-blocking. This can happen when another
process sets a tty stdin non-blocking because this setting persists on
the tty after that process exits (and the same fd is used for stdin,
stdout and stderr).